test(#4893): preserve queue guard lane coverage
Keep the moved stale-dispatch wiring regression under its established curated test module identity so the protected coverage ratchet remains exact.
